Cleanup color-scheme code to better match menu-collapse
