--- app/actions/color-scheme.ts
@@ -1,1 +1,0 @@
-export { action } from "~/modules/color-scheme/server";
