--- app/modules/color-scheme/types.ts
@@ -1,1 +1,0 @@
-export type ColorScheme = "dark" | "light" | "system";
