+++ app/routes.ts
@@ -4,7 +4,7 @@ const routes: RouteConfig = [
   index("pages/splash.tsx"),
   route("/brand", "pages/brand.tsx"),
   route("/healthcheck", "pages/healthcheck.tsx"),
-  route("/color-scheme", "actions/color-scheme.ts"),
+  route("/color-scheme", "actions/color-scheme/route.ts"),
 
   route("", "pages/docs-layout.tsx", { id: "docs" }, [
     route("/menu-collapse", "actions/menu-collapse/route.ts"),
